education_django/requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
coreapi==2.3.3
coreschema==0.0.4
croniter==0.3.31
Django==3.0.3
django-crispy-forms==1.9.0
django-debug-toolbar==2.2
django-filters==0.2.1
django-rq==2.3.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
entrypoints==0.3
flake8==3.7.9
idna==2.9
itypes==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
mccabe==0.6.1
openapi-codec==1.3.2
pycodestyle==2.5.0
pyflakes==2.1.1
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2019.3
redis==3.4.1
requests==2.23.0
rq==1.3.0
rq-scheduler==0.9.1
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.0
uritemplate==3.0.1
urllib3==1.25.8
